WIP module layout overhaul
